+++ codebase/apps/mdv_utils/src/MdvArithmetic/MdvArithmetic.cc
@@ -960,7 +960,7 @@ void MdvArithmetic::_remap(DsMdvx &mdvx)
 
 {
 
-  for (int ifld = 0; ifld < mdvx.getNFields(); ifld++) {
+  for (size_t ifld = 0; ifld < mdvx.getNFields(); ifld++) {
 
     MdvxField *field = mdvx.getField(ifld);
 
@@ -1118,7 +1118,7 @@ void MdvArithmetic::_autoRemapToLatLon(DsMdvx &mdvx)
 
 {
 
-  for (int ifld = 0; ifld < mdvx.getNFields(); ifld++) {
+  for (size_t ifld = 0; ifld < mdvx.getNFields(); ifld++) {
 
     MdvxField *field = mdvx.getField(ifld);
